+++ source/urc-service-vo/src/main/java/com/yks/urc/vo/PersonVO.java
@@ -17,7 +17,7 @@ public class PersonVO implements Serializable {
 
     private String personName;
 
-    private Byte gender;
+    private int gender;
 
     private String phoneNum;
 
@@ -77,11 +77,11 @@ public class PersonVO implements Serializable {
         this.personName = personName == null ? null : personName.trim();
     }
 
-    public Byte getGender() {
+    public int getGender() {
         return gender;
     }
 
-    public void setGender(Byte gender) {
+    public void setGender(int gender) {
         this.gender = gender;
     }
